% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-APA2_list.R
\name{APA2_list}
\alias{APA2_list}
\title{Formatierte Regressionstabelle}
\usage{
APA2_list(
  x,
  caption = "",
  note = "",
  output = stp25output::which_output(),
  custom.model.names = NULL,
  include.param = TRUE,
  include.gof = TRUE,
  include.custom = NULL,
  include.b = TRUE,
  include.se = TRUE,
  include.beta = FALSE,
  include.ci = FALSE,
  include.odds = FALSE,
  include.odds.ci = FALSE,
  include.statistic = FALSE,
  include.p = FALSE,
  include.stars = TRUE,
  include.df = FALSE,
  include.effects = c("ran_pars", "fixed"),
  conf.level = 0.95,
  conf.method = "Wald",
  digits = NULL,
  digits.param = 3,
  digits.odds = 2,
  digits.test = 2,
  digits.beta = 2,
  format = "fg",
  include.r = TRUE,
  include.pseudo = TRUE,
  include.rmse = TRUE,
  include.sigma = FALSE,
  include.variance = FALSE,
  include.devianze = FALSE,
  include.loglik = FALSE,
  include.test = FALSE,
  include.aic = TRUE,
  include.bic = include.aic,
  include.nobs = TRUE,
  rgroup = c("Parameter", "Goodness of fit"),
  test.my.fun = FALSE,
  dictionary = c(std.error = "SE", estimate = "b", p.value = "p"),
  col_names = NULL,
  ...
)
}
\arguments{
\item{x}{Regressionsobjekt}

\item{caption, note, output, custom.model.names, rgroup}{An Output}

\item{include.param, include.gof, include.custom}{Was soll ausgegeben werden}

\item{include.b, include.beta, include.ci, include.odds, include.se, include.statistic, include.odds.ci, include.p, include.stars}{an extract_coef}

\item{include.df, include.r, include.pseudo, include.rmse, include.sigma, include.variance, include.devianze, include.loglik, include.aic, include.bic, include.nobs, include.test}{An extract_gof()}

\item{include.effects, test.my.fun, conf.level, conf.method}{an extract_coef}

\item{digits.param, digits.odds, digits.test, digits.beta, formatan}{extract_coef}

\item{...}{nicht benutzt}
}
\value{
data.frame
}
\description{
Formatiert Listen mit Modellen zu Dataframs.
}
\examples{

fit1 <- lm(chol0 ~ rrs0 + med, hyper)
fit2 <- lm(chol0 ~ rrs0 + med + ak, hyper)
fit3 <- lm(chol0 ~ ak + med + rrs0 , hyper)
fit4 <- lmerTest::lmer(chol0 ~ rrs0 + med +  ak  +  (1 | g) , hyper)

coefs <- APA2_list(
  list(fit1,
       fit2,
       fit3, fit4),
  include.beta = TRUE,
  include.custom =
    data.frame(term = "M",  "A", "B", "C", "D")
)

# fit1 <- lm(chol0 ~  ak + rrs0 + med + g, hyper)
#
#
# x1 <- stp25stat:::APA2_list(
#   fits,
#   custom.model.names = c("lm", "glm", "lmer"),
#   digits = list(c(1, 2, 3, 4, 5, 6, 7),
#                 c(1, 2, 3, 4, 5, 6, 7),
#                 c(1, 2, 3, 4, 5, 6)),
#   include.custom = list(
#     Wald = c("F(1)=245",
#              "F(2)=241",
#              "F(3)=242"),
#     Chi = c("X(4)=2.45", "X(5)=24.5", "X(6)=24.5")
#   ),
#   include.pseudo = FALSE,
#   output = FALSE
# )

}
