% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC2.R, R/T-test.R, R/correlation.R,
%   R/describe-APA_x.R, R/describe-likert.R, R/describe-multiresponse.R,
%   R/dimred-pca.R, R/dimred-reliability.R, R/regression-rsquer.R, R/xtable.R
\name{APA_ICC}
\alias{APA_ICC}
\alias{APA_Ttest}
\alias{APA_Correlation}
\alias{Hmisc_rcorr}
\alias{APA_}
\alias{APA_NULL}
\alias{APA_Likert}
\alias{APA2_multiresponse}
\alias{APA_PCA}
\alias{APA_Reliability}
\alias{APA_R2}
\alias{APA_Xtabs}
\alias{APA_Xtabs.glm}
\alias{APA_Xtabs.formula}
\alias{APA_Xtabs.data.frame}
\alias{APA_Xtabs.default}
\title{APA Syle Table}
\usage{
APA_ICC(
  ...,
  caption = "ICC",
  note = "",
  type = c(1, 4),
  output = which_output()
)

APA_Ttest(
  x,
  data,
  caption = "T Test",
  note = alternative,
  output = stp25output::which_output(),
  var.equal = FALSE,
  paired = FALSE,
  alternative = "two.sided",
  include.mean = TRUE,
  include.d = TRUE,
  include.mean.diff = TRUE,
  include.se = FALSE,
  type = "t.test",
  digits = 2,
  ...
)

APA_Correlation(
  ...,
  caption = "Korrelation",
  note = "",
  output = which_output(),
  col_names = NULL,
  cor_diagonale_up = TRUE,
  type = c("pearson", "spearman"),
  include.mean = FALSE,
  include.n = TRUE,
  stars = TRUE,
  p.value = FALSE,
  include.stars = stars,
  include.p = p.value
)

Hmisc_rcorr(
  ...,
  cor_diagonale_up = TRUE,
  include.stars = TRUE,
  include.p = FALSE,
  include.mean = FALSE,
  include.n = TRUE,
  type = c("pearson", "spearman"),
  caption = "",
  note = ""
)

APA_NULL(x, ...)

APA_Likert(...)

APA2_multiresponse(
  Formula,
  data,
  caption = "",
  note = "",
  test = FALSE,
  order = FALSE,
  decreasing = TRUE,
  sig_test = "fischer.test",
  na.action = na.pass,
  use.level = 1,
  output = which_output(),
  ...
)

APA_PCA(
  data,
  ...,
  include.pca = TRUE,
  include.loading = include.pca,
  include.test = include.pca,
  include.plot = FALSE,
  include.kmo = TRUE,
  w = 5,
  h = 5,
  opene_graphic_device = FALSE,
  save_graphic = opene_graphic_device,
  caption = "Standardized loadings (pattern matrix) based upon correlation matrix",
  note = "",
  output = which_output(),
  main = ""
)

APA_Reliability(..., caption = "", note = "", output = which_output())

APA_R2(..., caption, note)

APA_Xtabs(x, ...)

\method{APA_Xtabs}{glm}(
  x,
  caption = "",
  output = stp25output::which_output(),
  thresh = 0.5,
  ...
)

\method{APA_Xtabs}{formula}(
  x,
  data = NULL,
  caption = "",
  output = stp25output::which_output(),
  labels = FALSE,
  addNA = FALSE,
  exclude = if (!addNA) c(NA, NaN),
  drop.unused.levels = FALSE,
  include.prop.chisq = TRUE,
  include.chisq = FALSE,
  include.correlation = FALSE,
  include.fisher = FALSE,
  include.mcnemar = FALSE,
  include.resid = FALSE,
  include.sresid = FALSE,
  include.sensitivity = FALSE,
  test = include.prop.chisq | include.chisq | include.fisher | include.mcnemar,
  ...
)

\method{APA_Xtabs}{data.frame}(
  data = NULL,
  formula,
  caption = "",
  output = stp25output::which_output(),
  labels = FALSE,
  ...
)

\method{APA_Xtabs}{default}(x, ...)
}
\arguments{
\item{...}{extra arguments}

\item{caption, note}{Ueberschrift an Output}

\item{type}{enweder "pearson" oder "spearman"}

\item{output}{Html oder Text}

\item{x}{An object to be converted into a tidy data.frame or Formula}

\item{data}{data.frame wenn x eine Formel ist}

\item{var.equal, paired, alternative}{an t.Test var.equal = FALSE  "two.sided"}

\item{include.mean}{Mittelwerte}

\item{include.d}{Cohens d}

\item{include.mean.diff}{Mittlere Differenzen}

\item{include.se}{Standardfehler noch nicht Fertig}

\item{digits}{Nachkommastellen}

\item{col_names}{nicht benutzt}

\item{cor_diagonale_up}{Diagonale oben oder unter}

\item{include.n}{Anzahl an gueltigen Werten}

\item{include.stars}{Sternchen als p-Werte}

\item{include.p}{Explizite p-Werte}

\item{Formula, test, order, decreasing, sig_test, na.action}{Interne Parameter}

\item{include.pca, include.loading, include.test, include.plot, include.kmo}{was soll ausgegeben werden}

\item{w, h, main, opene_graphic_device, save_graphic}{Fenster w=5 h=5 Ueberschrift}

\item{addNA, exclude, drop.unused.levels}{An xtabs() default = FALSE}
}
\value{
a data.frame or list with data.frame
}
\description{
APA_ICC:Intra-Klassen-Korrelation (mit psych).

\strong{APA_Ttest}:
 Berechnung von Mittelwerten und Mittelwertdifferenzen und des T-Test.

\strong{APA_Correlation}: Korrelationstabelle (Interkorrelationen mit der Hilfe
der Funktion  Hmisc::rcorr. Erlaubt ist die getrennte Auswertung ueber by=~b

Wird in APA2 verwendet Tabelle Arbeitet mit Multi2default()

APA_PCA: Faktoranalyse wie SPSS \link{PCA}

APA_R2: R2-Tabelle (noch nicht fretig)

\strong{Kreuztabellen} cross-tabulation, APA_Xtabs, APA2.xtabs :
Die Funktion Formatiert xtabs() mit NxM und NxMxO Tabellen.

\itemize{
 \item{"Haufigkeit/Prozent 1"}{
   Die Prozent werden ueber include.percent mit margin erstellt. Der Parameter
   add.margins wird automatisch vergeben.
   include.total, include.total.columns, include.total.sub, include.total.rows, include.percent,
   include.count.
   Feineinstellungen erfolgt ueber \code{margin = 2}).
   }
 \item{"Sensitivitaets Analyse"}{Nur bei 2x" Tabellen ueber test=TRUE}
 \item{"Sig.-Tests"}{Bei 2x" Tabellen Fischer sonst Chi-test.
 Die Berechnung erfolgt hier mit \link{assocstats}.
 Weiter Einstellungen sind Correlationen, Pearson, Kontingentkoeffizient
 berechnet alternativ steht auch  der Phi-Coefficient}

}
}
\examples{

 

 

APA_Ttest(m1+m2 ~ geschl, varana)
varanax<-Melt2(m1+m2~nr,varana , key="time", value="m")

# broom::tidy(with(varana, t.test(m1,m2 ) ))
# broom::tidy(  t.test(m~time, varanax, var.equal=FALSE)) )

APA_Ttest(m~time, varanax, paired = TRUE)

APA_Ttest(m~time, varanax, var.equal=TRUE)
APA_Ttest(m~time, varanax, var.equal=FALSE)
#broom::tidy(t.test(m~time, varanax, var.equal=TRUE))
#broom::tidy(t.test(m~time, varanax, var.equal=FALSE))

 

n <- 2 * 20
e <- rnorm(n)
dat <- stp25aggregate::Label(
  data.frame(
    a = rnorm(n) + e / 2,
    b = rnorm(n) + e,
    c = rnorm(n),
    d = rnorm(n) + e * 10,
    g = gl(2, 20, labels = c("Control", "Treat"))
  ),
  a = "Alpha",
  b = "Beta",
  c = "Gamma"
)


APA_Correlation( ~ a + b + c, dat)
APA_Correlation(a ~ c, dat)
APA_Correlation(a + b + c ~ d, dat)
APA_Correlation(a + b ~ c +  d, dat)
APA_Correlation(a + b + c ~ d, dat, groups = ~ g)
APA_Correlation( ~ a + b + c, dat)
APA_Correlation( ~ a + b + c, dat, include.p = TRUE)
APA_Correlation( ~ a + b + c, dat, include.p = TRUE, include.stars = FALSE)

 # library(arm)
 # windows(7,7)
 # corrplot(data[Cs(a,b,c,d)], abs=TRUE, n.col.legend=7)

 # install.packages("PerformanceAnalytics")

 #library("PerformanceAnalytics")
 #?chart.Correlation#(decathlon2.active[, 1:6], histogram=TRUE, pch=19)
 #data(managers)
 #chart.Correlation(managers[,1:8], histogram=TRUE, pch="+")


fit1<-lm(score ~ grade + treatment, schools)
fit2<-lm(score ~ grade + treatment + stdTest, schools)
APA_R2(fit1, fit2)

}
