% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-modell_pruefung.R
\name{APA_Autocorrelation}
\alias{APA_Autocorrelation}
\title{APA_Autocorrelation}
\usage{
APA_Autocorrelation(x, order = 1, ...)
}
\description{
Durbin–Watson d Test Der Durbin–Watson (DW) 
Test verwendet die geschaetzten Residuen um auf 
Autokorrelation erster Ordnung zu testen

Der DW d Test weist folgendes Intervall auf: 0<d<4
0:  Extrem positive Autokorrelation 
4:   Extrem negative Autokorrelation. 
2:   Keine Autokorrelation


Der Breusch–Godfrey (BG) Test ist flexibler,
 da er auf Autokorrelation höherer Ordnung testet und auf
}
