% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe-berechne-mean.R
\name{berechne}
\alias{berechne}
\alias{berechne.data.frame}
\alias{berechneMean}
\alias{berechne_all}
\alias{calc_median}
\alias{calc_mean}
\alias{Multi2default}
\alias{mean2}
\alias{median2}
\alias{sd2}
\alias{CI}
\alias{length2}
\alias{Mean2}
\alias{Mean2.formula}
\alias{Mean2.default}
\alias{Median2}
\alias{Median2.formula}
\alias{Median2.default}
\alias{Meanci2}
\alias{Prozent}
\alias{Prozent2default}
\alias{Prozent2APA}
\title{berechne: Berechne Mittelwerte}
\usage{
berechne(...)

\method{berechne}{data.frame}(data, ..., by = "1", type = 1,
  na.rm = TRUE, conf.interval = 0.95, .drop = TRUE)

berechneMean(data = NULL, measurevar, by = NULL, na.rm = TRUE,
  conf.interval = 0.95, .drop = TRUE)

berechne_all(data, x, by = "1", measure, type, fun = function(x)
  length(na.omit(x)), fm = NULL, digits = NULL, measure.name = NULL)

calc_median(x, digits = 2, n = length(x),
  median.style = get_my_options()$apa.style$mittelwert$median.style)

calc_mean(x, digits = 2, n = length(x),
  mean.style = get_my_options()$apa.style$mittelwert$mean.style)

Multi2default(x, digits = 0, n = length(x), use.level = 1,
  include.level = get_my_options()$apa.style$prozent$include_name)

mean2(x, na.rm = TRUE, ...)

median2(x, na.rm = TRUE, ...)

sd2(x, na.rm = TRUE, ...)

CI(x, ci = 0.95, na.rm = TRUE, ...)

length2(x, na.rm = FALSE)

Mean2(x, ...)

\method{Mean2}{formula}(x, data, ...)

\method{Mean2}{default}(x, digits = NULL, ...)

Median2(x, ...)

\method{Median2}{formula}(x, data, ...)

\method{Median2}{default}(x, digits = NULL,
  median.style = get_my_options()$apa.style$mittelwert$median.style, ...)

Meanci2(x, digits = NULL, ...)

Prozent(x, digits = 1, continuous = 3, breaks = NULL,
  labels = NULL, count_factor = c("yes", "ja", "T", "TRUE", 1),
  retur_tabel = FALSE)

Prozent2default(x, digits = 0, n = length(x), exclude = NA,
  max_factor_length = 25)

Prozent2APA(x, exclude = NA, digits = 1, max_factor_length = 35, ...)

Multi2default(x, digits = 0, n = length(x), use.level = 1,
  include.level = get_my_options()$apa.style$prozent$include_name)
}
\arguments{
\item{...}{alles weitere}

\item{data}{Objekt}

\item{by, measure, measure.name}{aus prepare Formula}

\item{type}{mean, median}

\item{na.rm}{NAs}

\item{conf.interval, ci}{Grenzen der Konfidenzintervalle CIs}

\item{.drop}{anplyr::ddply}

\item{measurevar}{Namen der Variablen}

\item{x}{measure.vars}

\item{fun}{Function an plyr::ddply}

\item{fm}{formel}

\item{digits}{Nachkommastellen}

\item{n}{intern}

\item{median.style, mean.style}{intern in calc_median}

\item{use.level}{welcher level wir gezaelt}

\item{include.level}{mit yes in labels?}

\item{continuous, breaks, labels}{fuer cut}

\item{count_factor}{für multi}

\item{retur_tabel}{intern}

\item{exclude}{geh noch nicht}

\item{max_factor_length}{lange Einträge kuerzen}
}
\value{
ein dataframe Objekt oder ein Character-String

Vector

Vector
}
\description{
Die Lagemasse werden ueber die Standard-Funktionen berechnet unterschied ist nur dass
Faktoren zu Zahlen transformiert werden und das \code{na.rm=TRUE} gesetzt wird.
CI = Hmisc::smean.cl.normal

Interne Funktion
berechne_all(data,x,...):
interne Funktion um Mittelwerte/Freq zu berechnen. Wird in Tabelle benutzt.

Prozent:
Interne Funktion um Prozent fuer die Tabellen zu berechnen.

Prozent2 wird in APA.formula errate_statistik2 verwendet
freq = Prozent2(x_NA, exclude, digits=digits.percent, rtrn = "df", max_factor_length = max_factor_length)
}
\examples{
# erlaubt:  varana \%>\% berechne(4, 5, by= ~geschl )
#  berechne(hyper, "chol0" )
#  names(hyper)
#  hyper \%>\% berechne(chol0,chol1,chol6,chol12, by=~med+g)
mean2(rnorm(100))
sd2(rnorm(100))
CI(rnorm(100))


Meanci2(c(1,1,1,1,2,2,4,4,5,5,5))

#set_my_options(prozent = list(digits = 1,
#style = 2,
#null_percent_sign= "."))
#Prozent2default(factor(c(1,2,3,3,3,5), 1:5))

}
