% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metcomp-bland_altman.R, R/metcomp-kappa.R
\name{BlandAltman}
\alias{BlandAltman}
\alias{BlandAltman.data.frame}
\alias{BlandAltman.formula}
\alias{Kappa}
\alias{Kappa2}
\title{BlandAltman}
\usage{
BlandAltman(.data, x, ...)

\method{BlandAltman}{data.frame}(.data, x, ...)

\method{BlandAltman}{formula}(x, .data, ...)

Kappa(xtb, ..., CI = FALSE)

Kappa2(xtb, ..., CI = FALSE, caption = "Cohen's Kappa-Koeffizient")
}
\arguments{
\item{x}{Objekt}

\item{...}{weitere Objekte nicht benutzt}
}
\value{
A \code{\link[tibble]{tibble}} with counted tagged NA values.
}
\description{
Diese Funktion ist ein Platzhalter
(see \code{\link[haven]{tagged_na}})
und zeigt die ...
}
\examples{
library(haven)


}
