% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-corr_plot.R
\name{corr_plot}
\alias{corr_plot}
\alias{corr_plot.formula}
\alias{corr_plot.lm}
\alias{corr_plot.data.frame}
\title{Correlation Plot}
\usage{
corr_plot(x, ...)

\method{corr_plot}{formula}(x, data, ...)

\method{corr_plot}{lm}(x, data = x$model, type = c("p", "r"),
  scales = list(x = list(relation = "free")), ylab = names(data)[1],
  xlab = "", layout = c(ncol(data) - 1, 1), ...)

\method{corr_plot}{data.frame}(data, jitter = FALSE, smooth = TRUE,
  lines = TRUE, pch = 20, digits = 2, cex.cor = NULL,
  method = "pearson", stars = FALSE, resize = FALSE, hist = TRUE,
  ...)
}
\arguments{
\item{x}{formula}

\item{...}{weitere Param an Lattice oder pairs}

\item{data}{a data matrix}

\item{type}{an xyplot c("p", "r")}

\item{scales}{an xyplot}

\item{ylab, xlab}{an xyplot}

\item{layout}{an xyplot}

\item{jitter}{Rauschen}

\item{smooth}{Gezeichnete Lineie}

\item{lines}{Regressinsgerade}

\item{pch}{Symbole  pch=20}

\item{digits}{Nachkommastellen in plot}

\item{cex.cor, resize}{Fixe groese mit cex.cor, resize abhaengig von r-Wert}

\item{method}{c("pearson", "kendall", "spearman")}

\item{stars}{Sternchen}

\item{hist}{Histogram TRUE/FLASE}

\item{data}{aus x$model}
}
\value{
nix
}
\description{
Function for making a correlation plot starting from a  formula and a data.frame
}
\examples{

# graphics.off()
# library(stpvers)

#windows(8,8)
corr_plot(~ m1 + m2 + m3 + m4, varana)
#windows(8,8)
corr_plot( m1~  m2 + m3 + m4, varana)

}
