% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-effsize.R
\name{Effsize}
\alias{Effsize}
\alias{APA_Effsize}
\alias{etaSquared2}
\alias{cohens.d}
\alias{cohens.d.default}
\alias{cohens.d.formula}
\title{Effsize}
\usage{
APA_Effsize(x, ...)

etaSquared2(x, type = 2, anova = FALSE, ...)

cohens.d(x, ...)

\method{cohens.d}{default}(x, y, ...)

\method{cohens.d}{formula}(x, data = list(), ...)
}
\arguments{
\item{x}{Objekt oder Formel}

\item{...}{weitere Optionen}

\item{type}{etaSquared2: Anova Type default ist 2}

\item{anova}{etaSquared2: Ausgabe der ANOVA Tabelle}

\item{x, y}{formel oder x, y}

\item{...}{}
}
\value{
Vector

vector
}
\description{
Measures of association
Pearson's r correlation Small 0.2,  Medium 0.5, Large 0.8
r2 coefficient of determination Small 0.04, Medium 0.25, Large 0.64
Quelle: http://www.ncbi.nlm.nih.gov/pmc/articles/PMC3444174/pdf/i1949-8357-4-3-279.pdf

Eta-Quadrat Kopie von lsr::etaSquared.
Die Ergebnisse entsprechen denen von SPSS (Univariat-Partial Eta Squared)

Cohen's d and Hedges g effect size
Between groups
Cohen's d Small 0.2, Medium 0.5, Large 0.8, Very large 1.3
Odds ratio (OR) Small 1.5, Medium 2, Large 3
Relative risk or risk ratio (RR) R Small 2, Medium 3, Large 4
Cohen, J. (1988).  Statistical power analysis for the behavioral sciences (2nd ed.).  New York:Academic Press.
}
\details{
Gestolen von https://cran.r-project.org/web/packages/effsize/effsize.pdf
}
\examples{

# etaSquared

fit1<-lm(y1~x1, anscombe)
#etaSquared2(aov (y1~x1, anscombe), anova=TRUE)
#etaSquared2(fit1, anova=TRUE)
etaSquared2(fit1 )


#require(stpvers)
set.seed(45)                        ## be reproducible
 x <- rnorm(10, 10, 1)
 y <- rnorm(10, 5, 5)

cohens.d(x, y)
varanax<-Melt2(m1+m2~nr,varana , key="time", value="m")
cohens.d(m~time, varanax )

}
