% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-extract.R
\name{extract_gof}
\alias{extract_gof}
\title{extract_gof}
\usage{
extract_gof(x, include.r = TRUE, include.pseudo = TRUE,
  include.rmse = TRUE, include.sigma = FALSE,
  include.variance = FALSE, include.devianze = FALSE,
  include.loglik = FALSE, include.test = FALSE, include.aic = TRUE,
  include.bic = include.aic, include.nobs = TRUE, digits = 2,
  fix_format = FALSE, ...)
}
\arguments{
\item{x}{objekt}

\item{include.r, include.pseudo, include.rmse, include.sigma, include.variance, include.devianze, include.loglik, include.aic, include.bic, include.nobs}{Alles wie bei APA_TAble}

\item{digits}{nachkommastellen}

\item{...}{extra Argumente}
}
\value{
tibble
}
\description{
extract_gof
}
\examples{


}
