% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-surv.R
\name{fix_format_p}
\alias{fix_format_p}
\title{fix_irgendwas fix_format_p ist nur hier in Verwendung
 fix_format_p sucht automatisch nach den p-Werten die meist an der Letzten stelle sind
 und gibt einen Vector-String mit der LAenge drei aus.
 Nicht zu verwechseln mit \code{rndr_P()}}
\usage{
fix_format_p(x, df1 = NULL, df2 = null, p = NULL)
}
\description{
fix_format_p Input(F, df,und p)  Output: (test, df, p.value)
}
