% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kano-plot.R
\name{kano_plot}
\alias{kano_plot}
\alias{kano_barchart}
\title{Kano Plot}
\usage{
kano_plot(
  x,
  main = "",
  xlim = c(0, 1),
  ylim = c(0, 1),
  mar = c(0, 1, 2, 1),
  legend.position = list(x = "right", y = NULL),
  legend.title = NULL,
  my.lines = "circle",
  col.bg = "gray95",
  col = NA,
  txt.bg = list(m = "M", i = "I", o = "O", a = "A"),
  cex.bg = 12,
  jitter = TRUE,
  cex = 1,
  cex.items = cex * 1,
  cex.lab = cex * 1.07,
  cex.legend = cex * 1.1,
  ylab = "Zufriedenheitsstiftung (CS+)",
  xlab = "Unzufriedenheitsstiftung (CS-)",
  center.axis = FALSE,
  use.labels = TRUE,
  use.total.strength = TRUE,
  use.categorie = TRUE,
  use.points = FALSE,
  ...
)

kano_barchart(
  x,
  data = x$molten,
  groups = x$groups[1],
  main = "Kano-Analyse",
  auto.key = list(space = "right"),
  prop.table = TRUE,
  ylab = if (prop.table) "Prozent" else "Anzahl",
  col = RColorBrewer::brewer.pal(6, "Dark2")[c(4, 1, 2, 3, 5, 6)],
  par.settings = list(superpose.polygon = list(col = col)),
  include.Q = TRUE,
  include.R = TRUE,
  include.n = TRUE,
  levels = c("M", "O", "A", "I", "R", "Q"),
  ...
)
}
\arguments{
\item{x}{Kano-Objekt}

\item{main, xlab, ylab, xlim, ylim}{an Plot}

\item{mar}{an par}

\item{legend.position, legend.title}{legende}

\item{my.lines}{Grafik Element Halbkreis fuer Indiferent}

\item{col.bg, txt.bg, cex.bg}{Hintergrund-Einstellung}

\item{col}{Farbe der Punkte bei barchart Farben RColorBrewer::brewer.pal(6, "Dark2")[c(4, 1, 2, 3, 5, 6)]}

\item{jitter}{Rauschen bei Ueberlappung}

\item{cex, cex.items, cex.lab, cex.legend}{Schriftgroesse}

\item{center.axis}{center.axis=FALSE}

\item{use.labels}{use.labels=FALSE  Zahlen oder Text als  Beschriftung bei den Streudiagrammen}

\item{use.total.strength}{use.total.strength=TRUE  groese der Schrift  als Indikator fuer Total Strenght}

\item{use.categorie}{use.categorie=TRUE  M, O, A,}

\item{use.points}{use.points = FALSE Punkte statt Text}

\item{...}{an lattice}

\item{data}{in kano_barchart molten data}

\item{groups}{in kano_barchart geht nicht}

\item{auto.key}{Key auf der rechten Seite}

\item{prop.table}{Prozent oder Anzahl}

\item{include.Q, include.R}{Q und R anzeigen}

\item{include.n}{Ueberschrift mit N=}

\item{levels}{Levels haendisch anordnen}
}
\value{
Tabelle als data.frame
}
\description{
Kano Plot
}
\examples{
#'
graphics.off()

# res1 <-  Kano( ~ . , DF[-c(1,2)])
res1$dysfunc
windows(9, 7)
kano_plot_del_bar(kano_res)
}
