% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe-likert.R
\name{Likert}
\alias{Likert}
\alias{brewer_pal_likert}
\alias{Likert.default}
\alias{Likert.formula}
\alias{print.likert}
\title{Likert}
\usage{
Likert(...)

brewer_pal_likert(
  n = 5,
  name = "RdBu",
  middle.color = "gray80",
  min_gray = 10,
  max_gray = 60
)

\method{Likert}{default}(data, ..., groups = NULL, labels = NULL, reverse.labels = FALSE)

\method{Likert}{formula}(x, data, labels = NULL, reverse.labels = FALSE)

\method{print}{likert}(x, ...)
}
\arguments{
\item{...}{Variablennamen}

\item{n, name, middle.color, min_gray, max_gray}{brewer_pal_likert Default 5 Stufen mit Grau in der Mitte}

\item{data}{Data.frame}

\item{groups}{(optional)}

\item{labels}{wenn die Labels anderst sein sollen}

\item{reverse.labels}{Codierung umderehen}

\item{x}{Objekt Formula}
}
\value{
a likert class list with the following elements:
  results: Ergebnisse der Haufigkeiten fuer zB Grafik
  names: Dataframe mit Namen
  freq: Data.frame mit nur den Haufigkeiten ohne NAs
  freq.na:Data.frame mit nur den Haufigkeiten mit NAs
  N: Stichprobengroesse als Zahl
  n: Anzahl der Gueltigen Werte je Item (Vektor)
  m: Mittelwerte (Vektor)
  sd: SD (Vektor)
  Mittelwert: M und SD Formatiert (Vektor)
  items: Daten
  grouping: data frame mit Gruppe (kann auch laenge Null haben)
  nlevels Anzahl an Levels
  levels: Labels fuer zb scale =list( at=nlevels, ...)
}
\description{
Analyse von Likertskalen.

Farb-Palette für Likert-plots die Funktion ist eine  eine Kopie von \code{brewer.pal.likert} kann aber auch schwarz-weiss.
und hat blau-rot als default
}
\examples{


set.seed(1)
n<-100
lvs<-c("--","-","o","+","++")
DF2<- data.frame(
  Magazines=gl(length(lvs),1,n,lvs),
  Comic.books=gl(length(lvs),2,n,lvs),
  Fiction=gl(length(lvs),3,n,lvs),
  Newspapers=gl(length(lvs),5,n,lvs))

DF2$Comic.books[sample.int(n/2)]<- lvs[length(lvs)]
DF2$Newspapers[sample.int(n/2)]<- lvs[1]
DF2$Magazines[sample.int(n/2)]<- lvs[2]

DF2<- transform(DF2, Geschlecht= cut( rnorm(n), 2, Hmisc::Cs(m, f)))
Res1 <- Likert(~., DF2[,-5] )
APA2(Res1)
APA2(Res2 <- Likert(.~ Geschlecht, DF2 ))

#require(HH)  # ?likertplot

#windows(7,3)
 #likertplot( Item   ~ .| Geschlecht , data=Res2$results,
 #            main = '',
 #            ylab = "",
 #            sub = "" ,
 #            xlab = "Prozent",
 #            col = brewer_pal_likert(5),
 #            rightAxis = FALSE,
 #            positive.order = TRUE,
 #            as.percent = TRUE,
 #            auto.key = list(space = "top", columns = 2) ,
 #            layout=c(2,2)
 #)




# data<- Res2$names
# data$mean<- Res2$m
#  barchart( Item~ mean |Geschlecht, Mymean2, xlim=c(1,5))
# windows(3,6)
# dotplot(Item ~ mean, data,
 #       groups=Geschlecht, xlim=c(.085, 5.15),
 #       type=c("p", "l"))


APA2(Res1, ReferenceZero=3)
APA2(Res1, ReferenceZero="-")
APA2(Res2, ReferenceZero=3, na.exclude=TRUE, type="freq")


brewer_pal_likert( ) #Default 5 Stufen mit Grau in der Mitte
#  HH::brewer.pal.likert(5, "RdBu", "gray80")

}
