% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{Metaanalysis}
\alias{Metaanalysis}
\title{Metaanalysis}
\usage{
Metaanalysis(x, output = which_output(), ...)
}
\arguments{
\item{x}{meta-Objekt}

\item{output}{an Autput}

\item{...}{an  APA2}
}
\value{
data.frame
}
\description{
Types Of Effects
An effect"could be almost any aggregate statistic of interest:
Mean, Mean difference, Mean change
Risk ratio, Odds ratio, Risk difference Incidence rate, Prevalence,
Proportion Correlation
}
\examples{

# library(meta)
# require(stpvers)
data2<- GetData("
                Nr            author  Ne    Me    Se Nc    Mc    Sc
                1    Blashki(75\%150)  13  6.40  5.40 18 11.40  9.60
                2     Hormazabal(86)  17 11.00  8.20 16 19.00  8.20
                3   Jacobson(75-100)  10 17.50  8.80  6 23.00  8.80
                4        Jenkins(75)   7 12.30  9.90  7 20.00 10.50
                5     Lecrubier(100)  73 15.70 10.60 73 18.70 10.60
                6        Murphy(100)  26  8.50 11.00 28 14.50 11.00
                7          Nandi(97)  17 25.50 24.00 10 53.20 11.20
                8      Petracca(100)  11  6.20  7.60 10 10.00  7.60
                9       Philipp(100) 105 -8.10  3.90 46 -8.50  5.20
                10     Rampello(100)  22 13.40  2.30 19 19.70  1.30
                11       Reifler(83)  13 12.50  7.60 15 12.50  7.60
                12       Rickels(70)  29  1.99  0.77 39  2.54  0.77
                13     Robertson(75)  13 11.00  8.20 13 15.00  8.20
                14      Rouillon(98)  78 15.80  6.80 71 17.10  7.20
                15           Tan(70)  23 -8.50  8.60 23 -8.30  6.00
                16 Tetreault(50-100)  11 51.90 18.50 11 74.30 18.50
                17      Thompson(75)  11  8.00  8.10 18 10.00  9.70")

print(meta::metacont(
  Ne,  Me,  Se,  Nc,  Mc,  Sc,
  sm = "SMD",
  data = data2,
  subset = 2
),
digits = 2)


mc1 <- meta::metacont(Ne, Me, Se, Nc, Mc, Sc,
                data=data1,
                studlab=author)
# round(c(mc1$TE.fixed, mc1$seTE.fixed^2), 4)
 #meta::forest(mc1)

APA2(mc1)


#Power calculations for the general linear model
#?pwr.f2.test
x <- pwr::pwr.f2.test(
  u = 6,
  f2 = .4,
  sig.level = 0.05,
  power = 0.80
)
res<- APA(x)

meta::metacont(n.e=10, mean.e=2.34, sd.e=1.14, 
n.c=20, mean.c=3.43, sd.c=1.01,  sm="SMD")
meta::metabin(10, 20, 15, 20, sm = "OR")

}
