% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetComp.R
\name{MetComp}
\alias{MetComp}
\alias{MetComp_Kappa}
\alias{MetComp_BAP}
\title{MetComp:  Uebereinstimmung und Praezision von Messwerten}
\usage{
MetComp(x, data, include.ci = TRUE, ci.level = 0.95, caption = NULL,
  note = "", output = stp25output::which_output())

MetComp_Kappa(x, include.ci = TRUE, ci.level = 0.95)

MetComp_BAP(..., X = NULL)
}
\arguments{
\item{x}{Formula Objekt}

\item{...}{weitere Objekte nicht benutzt}

\item{X}{Aufbereitete Daten aus Formula_Data}

\item{.data}{Daten}

\item{...}{an Formula_Data}
}
\value{
Ein bland_altman-Objekt mit den Daten (data) und der Statistik (stat).

list(stats, name, name.dif. met_A, met_B, groups)
}
\description{
Tukey Mean Difference oder auch Bland Altman Methode. Oft interessiert die Zuverlaessigkeit und Reproduzierbarkeit ein einer Diagnose. Die Beurteilung kann dabei durch einen Bewerter (Messverfahren) in wiederholter Form erfolgen und wird dann als Intra-Rater bezeichnet oder die Beurteilung eines Merkmals erfolgt durch mehrere Bewerter (Messverfahren). und hier spricht man von Inter-Rater.
Die Methode der Beurteilung der uebereinstimmung haengt von den jeweiligen Datentype ab.
Bei Nominalen wird abgezaehlt und die Rate der uebereinstimmung bewertet (Cohen-Koeffizient) Bei Ordinalen-Daten werden die gewichteten uebereinstimmungen ausgezaehlt (gewichteter Cohen-Koeffizient). Bei metrischen(stetigen) Daten werden die Differenzen beurteilt (Bland-Altman-Methode).
}
\details{
Bland-Altman-Methode Bias (d) systematische Abweichung Messfehler (s) Standardabweichung der Differenz Limits of agreement (LOA) Intervall von 95 (entspricht d+-1.96 -> es wird eine Normalverteilung unterstellt).
Methoden Die generische Funktion MetComp() kann sowohl Kappa als auch Tukey-means berechnen. Kappa kann aber auch ueber die xtab() und APA2 berechnet werden. Wobei hier nur 2x2-Tabellen untersucht werden und bei Kappa() sind hingegen auch mehrere ordinale Kategorien erlaubt sind.
aehnliche Methode ist ICC die aber eher zur Reliabilitaetsanalyse gehoert.
}
\examples{
   
#require(stp25stat)
#require(stp25plot)
#require(stp25output)
### Verschiedene Situationen Im folgenden habe ich eine fiktive Messung mit simulierten Daten

set.seed(0815)

n <- 100
DF <- data.frame(
  A = rnorm(n, 100, 50),
  B = rnorm(n, 100, 50),
  C = NA,
  D = NA,
  E = NA,
  F = NA,
  group = sample(gl(2, n / 2, labels = c("Control", "Treat")))
)

cutA <- mean(DF$A)
DF <- transform(
  DF,
  C = round(A + rnorm(n,-5, 20)),
  D = round(A + rnorm(n, 0, 10) + A / 10),
  E = A + ifelse(A < cutA, A / 5,-A / 5) + rnorm(n, 0, 10),
  F = A +  rnorm(n, 50, 10)
)



#### Methoden messen das Selbe

x<- MetComp(~A+C, DF)
#plot(x)
tab<- x$stat[,1:2]
names(tab)   <- c("Parameter", "Methoden messen das Selbe_M=0" )

x<- MetComp(~A+F, DF)
tab<- cbind(tab, "Methoden messen das Selbe_Fehler M=50"= x$stat$Unit)
#plot(x)



#### Methoden messen unterschiedlich Werte
x<- MetComp(~A+B, DF)
tab<- cbind(tab, "Methoden unterschiedliche_Fehler M=0"= x$stat$Unit)
#plot(x)




Output(tab, caption="BA" )

t1 <-
  APA2(with(DF, t.test( A, C,  paired = TRUE)), output=FALSE)
t2 <-
  APA2(with(DF, t.test( A, F,  paired = TRUE)), output=FALSE)
t3 <-
  APA2(with(DF, t.test( A, B,  paired = TRUE)), output=FALSE)


#Output(rbind( t1, t2, t3), caption="T-Test")


#### Methoden haben systematische Abweichungen

x<- MetComp(~A+D, DF)
#plot(x)



x<- MetComp(~A+E, DF)
#plot(x)


  ## ----sachs-627-data ---------------------------------------
Botulinum <- data.frame(
  A= factor(c(rep(1, 14), rep(1, 3),
              rep(0, 5),rep(0, 18)),
            1:0, c("+", "-")),
  B= factor(c(rep(1, 14), rep(0, 3),
              rep(1, 5),rep(0, 18)),
            1:0, c("+", "-")))
require(vcd)

Kappa(xtabs(~A+B, Botulinum))
data("SexualFun")
K <- Kappa(SexualFun)
#K
#confint(K)
#summary(K)
#print(K, CI = TRUE)

MetComp(~A+B, Botulinum)
MetComp_Kappa(SexualFun)



 
#- Understanding Bland Altman analysis
#Davide Giavarina
#Biochemia medica 2015;25(2) 141-51
#http://dx.doi.org/10.11613/BM.2015.015

set.seed(0815)
DF<- data.frame(
  A=c(1, 5,10,20,50,40,50,60,70,80, 90,100,150,200,250,300,350,400,450,500,550,600,650,700,750,800,850,900, 950,1000),
  B=c(8,16,30,14,39,54,40,68,72,62,122, 80,181,259,275,380,320,434,479,587,626,648,738,766,793,851,871,957,1001, 980),
  group= sample(gl(2, 15, labels = c("Control", "Treat")))
)

MetComp2(~A+B, DF, caption = "Giavarina")


}
