% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-polr.R, R/z-tidy.R
\name{Ordnen.polr}
\alias{Ordnen.polr}
\alias{Ordnen}
\alias{Ordnen.default}
\alias{Ordnen.anova}
\alias{Ordnen.aov}
\alias{Ordnen.lm}
\alias{Ordnen.glm}
\alias{Ordnen.merModLmerTest}
\alias{Ordnen.lmerModLmerTest}
\title{Ordnen (Tidy)}
\usage{
\method{Ordnen}{polr}(x, digits = 2, include.b = TRUE,
  include.se = TRUE, include.ci = FALSE, include.odds = TRUE,
  test.my.fun = FALSE, ...)

Ordnen(x, ...)

\method{Ordnen}{default}(x, test.my.fun = FALSE, ...)

\method{Ordnen}{anova}(x, ...)

\method{Ordnen}{aov}(x, include.eta = TRUE, include.sumsq = TRUE,
  include.meansq = FALSE, test.my.fun = FALSE, ...)

\method{Ordnen}{lm}(x, include.b = TRUE, include.se = TRUE,
  include.beta = FALSE, include.ci = FALSE, include.r = TRUE,
  include.ftest = FALSE, ci.level = 0.95, test.my.fun = FALSE, ...)

\method{Ordnen}{glm}(x, include.b = TRUE, include.se = TRUE,
  include.ci = FALSE, include.odds = TRUE,
  include.odds.ci = include.ci, include.rr = FALSE,
  include.rr.ci = include.ci, include.b.ci = include.ci,
  include.test = "lrt", include.r = TRUE, include.pseudo = include.r,
  include.loglik = TRUE, ci.level = 0.95, digits = 2,
  test.my.fun = FALSE, ...)

\method{Ordnen}{merModLmerTest}(x, include.b = TRUE, include.se = TRUE,
  include.beta = FALSE, include.ci = FALSE, include.odds = FALSE,
  ci.level = 0.95, ...)

\method{Ordnen}{lmerModLmerTest}(x, include.b = TRUE,
  include.se = TRUE, include.beta = FALSE, include.ci = FALSE,
  include.odds = FALSE, ci.level = 0.95, test.my.fun = FALSE, ...)
}
\arguments{
\item{x}{Objekt}

\item{include.b}{Estimate}

\item{include.se}{SE Standardfehler}

\item{include.ci, ci.level}{95-CI mit ci-Level}

\item{...}{weitere Objekte nicht benutzt}

\item{include.eta}{Eta Quadrat}

\item{include.sumsq, include.meansq}{Quadrat- Summen}

\item{include.beta}{standartisiertes beta}

\item{include.b.ci, include.odds, include.rr.ci}{95 Konfidenzintervalle}

\item{rr}{RR Relatives Risiko}
}
\value{
ein data.frame-Ojekt oder eine Liste von data.frames. Im Attribut N sind die Stichprobengroesse
und notes
}
\description{
Diese Funktion meine Version von tidy.

ANOVA - Methode ueber broom::tidy

Regression - Methode ueber basr::summary (lm und glm)

lmerTest::lmer returns an object of class 'lmerModLmerTest' (previously 'merModLmerTest')
to clarify that 'lmerModLmerTest' extends  'lmerMod' – not 'merMod'. The merMod class includes generalized
and nonlinear mixed models and lmerTest is only designed for linear mixed models.
}
