% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred-pca.R
\name{PCA}
\alias{PCA}
\alias{Principal2}
\alias{PCA2}
\alias{Principal}
\alias{PCA.default}
\title{PCA (Faktoranalyse)}
\usage{
PCA(x, ...)

Principal2(x, ...)

PCA2(x, ...)

Principal(x, ...)

\method{PCA}{default}(
  data,
  nfactors = 1,
  cut = 0.35,
  residuals = FALSE,
  rotate = "varimax",
  n.obs = NA,
  covar = FALSE,
  scores = TRUE,
  missing = FALSE,
  impute = "median",
  oblique.scores = TRUE,
  method = "regression",
  sort = TRUE,
  N = nrow(data),
  caption = "Standardized loadings (pattern matrix) based upon correlation matrix",
  ...
)
}
\arguments{
\item{...}{weitere Argumente wie:
digits = 2 Number of digits to use in printing
all = FALSE if all=TRUE, then the object is declassed and all output from the function is printed
cut = NULL Cluster loadings < cut will not be printed. For the factor analysis functions (fa and factor.pa etc.), cut defaults to 0, for ICLUST to .3, for omega to .2.
sort = TRUE Cluster loadings are in sorted order}

\item{data}{Daten als data.frame}

\item{nfactors}{Anzahl Faktoren default  nfactors=1}

\item{cut}{Cluster loadings  cut= .30,     <  will not be printed.}

\item{residuals}{Anzeige der Residuen  default: FALSE}

\item{rotate}{Rotation default = "varimax"}

\item{n.obs}{default= NA}

\item{covar}{default= FALSE}

\item{scores}{default= TRUE}

\item{missing}{default= FALSE}

\item{impute}{Imputation default = "median" NA ist keine Imputation}

\item{oblique.scores}{default= TRUE}

\item{method}{default= "regression"}

\item{sort}{default= TRUE}

\item{N}{ist nicht zu ändern}
}
\value{
HTML oder Test Output
}
\description{
Gestohlen von psych::print.psych.fa \link{print.psych}.
Die Ergebnisse stimmen mit SPSS weitgehend ueberein. Ueberprueft habe ich das
anhand der Daten vom Stampfl.
mehr unter \link{fkv}

Principal() ist die Kopie von principal aus psych
}
\details{
Interpretation:

KorrelationsMatrix: Korrelation von mindestens r=.30

Erklaerte Gesamtvarianz (Eigenwerte)
Erklärte Gesamtvarianz => Kummulierte %



Measures of Appropriateness of Factor Analysis
KMO- und Bartlett-Test (partiellen Korrelationen zwischen Itempaaren)
KMO: Minimum   0.50

Bartlett's test of sphericity (überprueft die Nullhypothese, ob die Korrelationsmatrix eine Identitaetsmatrix ist.)
gut ist wenn p < .050

Anzahl an Faktoren
Faustregel nur Eigenwerte größer als eins (Kaiser-Guttman-Kriterium)
oder Scree-Plot hier wird der charakteristische Knick als kriterium gewaehlt


Rotierte Komponentenmatrix (varimax)


Quelle: \url{http://statistikguru.de/spss/hauptkomponentenanalyse/auswerten-und-berichten.html}

Deutsch

Es wurde eine Hauptkomponentenanalyse durchgeführt, um die wichtigsten,
unabhängigen Faktoren zu extrahieren. Das Kaiser-Meyer-Olkin-Kriterium war .867 und der Bartlett-Test hochsignifikant (p < .001), was eine ausreichend hohe Korrelation zwischen Items darstellt, um eine Hauptkomponentenanalyse durchzuführen. Nur Faktoren mit Eigenwerten ≥ 1 wurden in Betracht gezogen (Guttman, 1954; Kaiser, 1960).
Eine Überprüfung des Kaiser–Kriteriums und Scree-Plots rechtfertigte
die Extraktion von zwei Faktoren, jeweils mit Eigenwerten über 1,
die eine Gesamtvarianz von 60.45 aufklären. Unter den Lösungen lieferte die Varimax rotierte zweifaktor-Lösung die Lösung, die am besten zu interpretieren war, bei der die meisten Items nur auf einen der beiden Faktoren hohe Ladungen zeigten.

English

We performed a Principal Component Analysis (PCA) to extract the most important independent factors. The Kaiser–Meyer–Olkin measure of sampling adequacy was .867, representing a relatively good factor analysis, and Bartlett’s test of Sphericity was significant (p < .001), indicating that correlations between items were sufficiently large for performing a PCA. Only factors with eigenvalues ≥ 1 were considered (Guttman, 1954; Kaiser, 1960).
Examination of Kaiser’s criteria and the scree-plot yielded empirical justification for retaining two factors with eigenvalues exceeding 1 which accounted for 60.45 of the total variance. Among the factor solutions, the varimax-rotated two-factor solution yielded the most interpretable solution, and most items loaded highly on only one of the two factors.
}
\examples{

pc <- psych::principal(Harman74.cor$cov,4,rotate="varimax")
mr <- psych::fa(Harman74.cor$cov,4,rotate="varimax")  #minres factor analysis
pa <- psych::fa(Harman74.cor$cov,4,rotate="varimax",fm="pa")  # principal axis factor analysis

round(psych::factor.congruence(list(pc,mr,pa)),2)

APA2(pc)
APA2(mr)
APA2(pa)


# -- Buehl Faktoranalyse Seite 475
## some(DF <- GetData("Raw data/FKV.SAV"))
# -  Freiburger Fragebogen zur Krankheitsverarbeitung
# #DF <- DF[,-1]
# APA2( ~., DF, test=T)
# library(arm)
# windows(5,5)
# corrplot(DF, abs=TRUE, n.col.legend=7)#  corrplot {arm}
# SaveData( )
# APA_PCA(DF, 5, cut=.35,include.plot = FALSE)



}
