% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-prepare_output.R
\name{prepare_output}
\alias{prepare_output}
\title{prepare_output}
\usage{
prepare_output(x, caption = NULL, note = NULL, N = NULL, labels = NA, ...)
}
\arguments{
\item{x}{Objekt Vector oder auch Formel}

\item{caption}{bei verwendung von Formeln}

\item{note}{Fehlende Werte}

\item{N}{Dezimalstellen bei zB Mean2}

\item{labels}{Grenzen der Konfidenzintervalle}

\item{...}{Weitere Argumente}
}
\value{
data.frame mit attributen
}
\description{
Ergebnis fuer Output vorbereiten
}
