% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable-prop_test.R
\name{Prop_Test}
\alias{Prop_Test}
\alias{Prop_Test2}
\alias{Prop_Test.default}
\title{Prop Test}
\usage{
Prop_Test(x, ...)

Prop_Test2(x, ..., caption = "", note = "", output = which_output())

\method{Prop_Test}{default}(
  x,
  p = 1/nlevels(x),
  n = length(na.omit(x)),
  conf.level = 0.95,
  digits = 0,
  include.ci = TRUE,
  ...
)
}
\arguments{
\item{x}{data.frame-Objekt}

\item{p}{Probability 0.50}

\item{n}{anzahl}

\item{conf.level}{Conf-Intervall 95}

\item{digits}{Nachkomnmastellen}
}
\value{
data.frame
}
\description{
Ausgabe von Konfidenzintervallen von Haufikeiten
}
\examples{
set.seed(234)
n<-999
library(BayesianFirstAid)
smokers  <- c( 83, 90, 129, 70 )
patients <- c( 86, 93, 136, 82 )


g =gl(3, n/3, labels = c("Control", "Treat A", "Treat B"))
g2<- g[sample.int(n)]
levels(g2)<- c("male", "female", "female")
data<- data.frame(g=g, g2=g2,
                  x=rnorm(n) )[sample.int(n)[1:78],]
#Prop_Test2(data$g)
x<-as.data.frame(table(data$g))$Freq

APA2(~g+g2, data, type="freq.ci")
APA2(g~g2, data, type="freq.ci")
}
