% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rndr.R
\name{rndr_CI2}
\alias{rndr_CI2}
\title{rndr_CI}
\usage{
rndr_CI2(ci, digits = 3, sep = ",", sep_1 = "[", sep_2 = "]",
  format = "g")
}
\arguments{
\item{ci}{matrix}

\item{digits}{Nachkommastellen}

\item{sep, sep_1, sep_2}{Seperator}

\item{format}{an formatC "fg"}
}
\value{
string
}
\description{
Formatieren von Konfidenzintervallen.
}
