% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-surv.R
\name{rndr_percent_CI}
\alias{rndr_percent_CI}
\title{rendere percent and CIs}
\usage{
rndr_percent_CI(x, low, upr, digits = 0, prc = "\% ", sep = ", ",
  sep_1 = "[", sep_2 = "]")
}
\arguments{
\item{x, low, upr}{Wert, Lower und Upper}

\item{digits}{Nachkommastellen}

\item{prc, sep, sep_1, sep_2}{Formatierungs-Symbole}
}
\value{
string
}
\description{
rendere percent and CIs
}
