% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred-sem.R
\name{SEM}
\alias{SEM}
\alias{SEM.default}
\alias{APA2.lavaan}
\title{SEM}
\usage{
SEM(x, ...)

\method{SEM}{default}(x, ...)

\method{APA2}{lavaan}(fit, baseline.model = NULL, caption = "",
  note = "", type = "all", ci = FALSE, est = FALSE, ...)
}
\arguments{
\item{x}{Objekt}

\item{...}{weitere Objekte nicht benutzt}
}
\value{
Ein lavaan -Objekt oder html.
}
\description{
Struckturgleichungs Modelle mit lavaan. Die Funktionen sind Kopieen von lavaan
und helfen den Output zu erstellen.
SEM ist davei einfach lavaan::sem(x, ...)
mehr unter \link{fkv}

\subsection{APA2.lavaan}{Ueber APA2 wird die Ausgabe formatiert. Ausgegeben werden die Guetemasse und der ModelFit.
Loading ist dabei der standartisierte Estimate und Communality ist die quadrierte Ladung

}

Chi-Quadrat-Wert

ML:  Validitaet des Models H0: empirische Kovarianz entspricht modelltheoretischer Kovarianz
Chi-Quadrat/df moeglichst klein (Chi-Quadrat/df<2.5 oder p<0.100)
Ist nur zuverlaessig wenn Normalverteilung und ausreichend grosse Stichprobe gegeben ist.

Model test Baseline model Chi-Quadrat Null-Modell wenn signifikant dann besteht die Gefahr einer Fehl-Spezifikation

Goodness-of-Fit-Index (GFI)

Ist vergleichbar mit dem Bestimmtheitsmass in der Regressionsanalyse, also ein Mass fuer die erklaerende Varianz
GFI>0.90
Adjusted-Goodness-of-Fit-Index (AGFI)
Analog wie GFI nur korrigiert durch df und Anzahl an Variablen
AGFI>0.90

Normed-Fit-Index NFI

Vergleicht das Modell mit einem Model bei dem alle Manifesten Variablen un-korreliert angenommen werden
NFI>0.90
Comparative-Fit-Index

Wie NFI nur korrigiert durch df und Anzahl an Variablen
CFI>0.90

Root-Mean-Square-Error of Approximation (RMSEA)
RMSEA<0.05

Backhaus Multivariate Analysemethoden 11 AuflageSeite 383
Moosbrugger, Kelava 2012 Testtheorie 2. Auflage Seite 339
}
\examples{
head(fkv)
APA2( ~., fkv, test=T)
library(arm)
windows(5,5)
corrplot(fkv, abs=TRUE, n.col.legend=7)#  corrplot {arm}
SaveData( )
Principal2(fkv, 5, cut=.35)

library(lavaan)
library(semPlot)


Model<-'
Verarbeitung =~ F5+F16+F22+F9+F26+F6+F35+F33+F12+F34+F4
Coping =~ F7+F8+F17+F14+F15+F18+F19+F1+F13+F20
Vertrauen =~ F28+F27+F31+F29
Religion =~F21+F25+F30+F23+F24
Distanz =~F3+F2+F10+F11

'
fit.Lavaan <- sem( Model, data=fkv)
APA2(fit.Lavaan)
# parameterEstimates(fit.Lavaan)
# Est <- parameterEstimates(fit.Lavaan, ci = FALSE, standardized = TRUE)
# #fitMeasures(fit.Lavaan, c("chisq", "df", "pvalue", "cfi", "rmsea"))
# #round( inspect(fit.Lavaan,"r2")  ,2)
# #parTable(fit.Lavaan)
# #show(fit.Lavaan)
# anova(fit.Lavaan)

semPaths(fit.Lavaan, "std", rotation=2, title = FALSE)
title("Std", line = 3)
}
