% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-stp25stat.R
\docType{package}
\name{stp25stat-package}
\alias{stp25stat}
\alias{stp25stat-package}
\title{stp25stat: Functions for Statistical Computations}
\description{
Convert statistical analysis objects from R into APA-Tabls.
}
\author{
\strong{Maintainer}: Wolfgang Peter \email{w.peter@statistik-peter.at}

}
