% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-extract.R
\name{tidy_lmer}
\alias{tidy_lmer}
\title{Fix tidy_lmer
https://github.com/tidymodels/broom/issues/309}
\usage{
tidy_lmer(x, effects = c("ran_pars", "fixed"), scales = NULL,
  ran_prefix = NULL, conf.int = FALSE, conf.level = 0.95,
  conf.method = "Wald", ...)
}
\arguments{
\item{x}{Objekt}

\item{effects, scales, ran_prefix, conf.int, conf.level, conf.method}{Param antidy}

\item{...}{extra params}
}
\value{
data.frame tibble
}
\description{
Fix tidy_lmer
https://github.com/tidymodels/broom/issues/309
}
