% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-vif.R
\name{VIF}
\alias{VIF}
\alias{VIF.default}
\alias{VIF2}
\alias{VIF.merModLmerTest}
\title{variance inflation factor}
\usage{
VIF(x, ...)

\method{VIF}{default}(x, ...)

VIF2(..., caption = "VIF", notes = "")

\method{VIF}{merModLmerTest}(x, ...)
}
\arguments{
\item{x}{fit-Objekt}

\item{...}{weitere Einstellungen}

\item{caption, note}{Ueberschrift an Output}
}
\description{
VIF  values over 5 are troubling, should probably investigate anything over 2.5.

VIF.default lm Methode aus car::vif

VIF.merModLmerTest Quelle https://hlplab.wordpress.com/2011/02/24/diagnosing-collinearity-in-lme4/
}
\examples{

\dontrun{
 fit<-lm(prestige ~ income + education, data=Duncan)
 car::vif(fit)
 VIF(fit)
}

}
\author{
Wolfgang Peter
}
