% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetComp.R, R/T-test.R, R/correlation.R,
%   R/describe-APA.R, R/dimred-reliability.R, R/regression-APA-lm.R,
%   R/regression-polr.R, R/regression-surv.R, R/xtable.R
\name{APA.bland_altman}
\alias{APA.bland_altman}
\alias{APA.htest}
\alias{APA.ScalarIndependenceTest}
\alias{APA.biVar}
\alias{APA}
\alias{APA.NULL}
\alias{APA.default}
\alias{APA.formula}
\alias{APA.stp25_reliability}
\alias{APA.lm}
\alias{APA.glm}
\alias{APA.polr}
\alias{APA.survfit}
\alias{APA.survdiff}
\alias{APA.xtabs}
\alias{APA.table}
\title{APA Style Text-Ausgabe}
\usage{
\method{APA}{bland_altman}(x, ...)

\method{APA}{htest}(x, ...)

\method{APA}{ScalarIndependenceTest}(x, ...)

\method{APA}{biVar}(x, ...)

APA(x, ...)

\method{APA}{NULL}(x, ...)

\method{APA}{default}(x, ...)

\method{APA}{formula}(x, data, exclude = NA, max_factor_length = 25,
  ...)

\method{APA}{stp25_reliability}(x, ...)

\method{APA}{lm}(x, include.r = TRUE)

\method{APA}{glm}(x, ...)

\method{APA}{polr}(x, ...)

\method{APA}{survfit}(x, ...)

\method{APA}{survdiff}(x)

\method{APA}{xtabs}(x, ...)

\method{APA}{table}(x, ...)
}
\arguments{
\item{x}{Objekt fit, formula usw}

\item{...}{weitere Objekte}

\item{data}{data.frame}

\item{exclude, max_factor_length}{an Prozent2default}

\item{include.r}{APA.lm: R-Squar}
}
\value{
Character Vector mit einem oder meheren Eintraegen
}
\description{
APA.htest: T-Test

APA.biVar: Korrelation mit Hmisc::spearman2

Ausgabe von APA-Style formatiertem Text.

APA.lm F-Test aus lm und Anova

APA.survfit  Mediane berechnen.

APA.survdiff Log-Rank-Test  berechnen.
}
\examples{
 # T-Test
require(coin)
APA(coin::wilcox_test(mpg ~ factor(vs), mtcars))
APA(wilcox.test(mpg ~ vs, mtcars))
APA(t.test(mpg ~ vs, mtcars))

# Correlation
APA(Hmisc::spearman2(mpg ~ factor(vs), mtcars))
 ## Hmisc::spearman2 Wilkox-Test

APA(mpg ~ cyl, mtcars)
APA(glm(vs ~ mpg, mtcars, family = binomial()))
APA(lm(mpg ~ drat + wt + qsec, mtcars))
APA(aov(mpg ~ drat + wt + qsec, mtcars))

\dontrun{

 APA( ~ rrs0 + rrs1 + g, hyper)
 APA(chol0+chol1 ~ g, hyper) das geht nicht
 }

}
