% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-modell_pruefung.R,
%   R/regression-vif.R
\name{APA_Validation}
\alias{APA_Validation}
\alias{APA_Multicollinearity}
\alias{APA_Autocorrelation}
\alias{APA_Heteroscedasticity}
\alias{APA_Durbin_Watson}
\alias{APA2.durbinWatsonTest}
\alias{VIF}
\alias{VIF.default}
\alias{APA_vif}
\alias{VIF2}
\alias{VIF.merModLmerTest}
\title{GOF 
 
 
Im Folgenden die wichtige Voraussetzungen der Regressionsanalyse}
\usage{
APA_Validation(..., include.ftest = TRUE, include.loglik = FALSE,
  include.minus.LL = include.loglik, include.pseudo = TRUE,
  include.r = include.pseudo, include.heteroskedasticity = TRUE,
  include.durbin = TRUE, include.levene = FALSE,
  include.bartlett = FALSE, include.multicollin = FALSE,
  include.vif = FALSE, include.sigma = FALSE, include.rmse = FALSE,
  include.aic = TRUE, include.bic = include.aic,
  include.residual = TRUE, include.normality = TRUE,
  include.deviance = TRUE, caption = "Testing Regression Models",
  note = "", names = NULL)

APA_Multicollinearity(x, caption = "Test for Multicollinearity ",
  note = "", include.vif = TRUE, include.cor = TRUE,
  output = which_output())

APA_Autocorrelation(x, order = 1, ...)

APA_Heteroscedasticity(x...)

APA_Durbin_Watson(x,
  caption = "Durbin-Watson Test for Autocorrelated Errors",
  note = NULL, ...)

\method{APA2}{durbinWatsonTest}(x,
  caption = "Durbin-Watson Test for Autocorrelated Errors",
  note = NULL, ...)

VIF(x, ...)

\method{VIF}{default}(x, ...)

APA_vif(..., caption = "VIF", notes = "")

VIF2(..., caption = "VIF", notes = "")

\method{VIF}{merModLmerTest}(x, ...)
}
\arguments{
\item{...}{weitere Einstellungen}

\item{include.ftest, include.loglik}{F-sratistik}

\item{include.r, include.pseudo}{R-Quadrat}

\item{include.heteroskedasticity}{Breusch-Pagan test}

\item{include.durbin}{autocorrelation}

\item{include.levene}{homogeneity of variance across groupssiehe T-Test}

\item{include.bartlett}{Homogeneity of Variances siehe T-Test}

\item{include.vif}{noch nicht Implementiert}

\item{include.sigma, include.rmse}{RMSE Extract Residual Standard Deviation Sigma}

\item{caption, note}{Ueberschrift an Output}

\item{x}{model- fit}

\item{include.cor}{Bivariate Correlations}

\item{x}{fit-Objekt}
}
\description{
Das Modell ist linear in den Parametern.
Die Residuen sind normalverteilt und im Mittel sind die Residuen null.
Die Zahl der Beobachtungen muss größer sein als die Zahl der zu schätzenden Parameter: n > k.
 
\strong{APA_Validation}
Validation of Linear Models Assumptions
(Testing Linear Regression Models)

Multi-collinearity diagnostics

Conducts a series of checks for multicollinearity.


Bei perfekter Multikollinearität lässt sich das Modell nicht lösen.

APA_Autocorrelation: 
Der Durbi-Watson (DW) 
Test verwendet die geschaetzten Residuen um auf 
Autokorrelation erster Ordnung zu testen.

Der DW d Test weist folgendes Intervall auf: 0<d<4
0:  Extrem positive Autokorrelation 
4:  Extrem negative Autokorrelation. 
2:  Keine Autokorrelation


Der Breusch-Godfrey (BG) Test ist flexibler,
da er auf Autokorrelation hoeherer Ordnung testet.

APA_Heteroscedasticity: mit (lmtest::bptest) Breusch-Pagan test against heteroskedasticity

APA_Durbin_Watson:  APA_Durbin_Watson(fit, max.lag=1, simulate=TRUE, reps=1000,
method=c("resample","normal"),
alternative=c("two.sided", "positive", "negative")): Durbin-Watson Test for Autocorrelated Errors. 
  Kopie der Funktion car::durbinWatsonTest

APA2.durbinWatsonTest: Methode für car::durbinWatsonTest Kopie von car:::print.durbinWatsonTest

multicollinearity diagnostic information 

VIF (variance inflation factor)

 values over 5 are troubling, should probably investigate anything over 2.5.
 
 Fuer Squareds und Correlations design matrix  \code{rockchalk::mcDiagnose(fit)}

VIF.default lm Methode aus car::vif

APA_vif:  variance inflation factor, aka VIF

VIF.merModLmerTest Quelle https://hlplab.wordpress.com/2011/02/24/diagnosing-collinearity-in-lme4/
}
\details{
\itemize{
\item{loglik} Likelihood Ratio Test The log-likelihood from the intercept-only restricted model.
-2LL: The LL (log-likelihood from the fitted model)
llhNull	(The log-likelihood from the intercept-only restricted model),
G2	(Minus two times the difference in the log-likelihoods)

\item{Autocorrelation} Durbin-Watson test for autocorrelation of disturbances. Ist nur bei Zeitreihendaten sinnvoll.

\item{Homogeneity of Variances} Levene Computes Levene's test for homogeneity of variance across groups.
 Bartlett Test of Homogeneity of Variances

\item{Heteroskedasticity} Breusch-Pagan test against heteroskedasticity.

\item{VIF} variance inflation factor. VIF values over 5 are troubling, should probably investigate anything over 2.5.

\item{Residual} RMSE values should be low (<0.5 and <0.3, respectively).
 SigmaResidual standard error  When the residual standard error is exactly 0 then the model fits the data perfectly (likely due to overfitting)
\item{ R-Quadrats } Cox und Snell R2: [ 0.2 = akzeptabel, 0.4 = gut ] 
Nagelkerke R2: [ 0.2 = akzeptabel, 0.4 = gut, 0.5 = sehr gut] 
McFaddens R2: [ 0.2 = akzeptabel, 0.4 = gut ] (see pR2)
}
}
\examples{

hkarz$Lai <- factor(hkarz$lai, 0:1, Cs(.neg, .pos))
fit2 <- glm(gruppe ~ tzell + Lai, hkarz, family = binomial)
fit1<-lm(score ~ grade + treatment + stdTest, schools)
APA_Validation(x) 
APA_Validation(fit1,fit2, include.pseudo = TRUE, include.r = TRUE, include.loglik = TRUE,
               include.rmse = TRUE)
               
               

# APA_Durbin_Watson
x<-lm(score ~ grade + treatment + stdTest, schools)
APA2(car::durbinWatsonTest(x))
DW_Test2(x)

lmtest::dwtest(x)
car::durbinWatsonTest(x)


\dontrun{
 fit<-lm(prestige ~ income + education, data=Duncan)
 car::vif(fit)
 VIF(fit)
}

}
