% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-effsize.R
\name{Effsize}
\alias{Effsize}
\alias{APA_Effsize}
\alias{APA2.emmGrid}
\alias{APA2.emm_list}
\alias{APA2.visreg}
\alias{APA2.eff}
\alias{APA2.efflist}
\alias{etaSquared2}
\alias{cohens.d}
\alias{cohens.d.default}
\alias{cohens.d.formula}
\title{Effsize}
\usage{
APA_Effsize(x, ...)

\method{APA2}{emmGrid}(
  x,
  caption = "Estimated marginal means",
  note = attr(x, "mesg"),
  output = which_output()
)

\method{APA2}{emm_list}(
  x,
  caption = "Estimated marginal means",
  note = attr(x, "mesg"),
  output = which_output(),
  include.means = TRUE,
  include.contrasts = TRUE,
  digits = 2
)

\method{APA2}{visreg}(x, caption = x$meta$y, note = "", include.ci = TRUE, digits = 2, ...)

\method{APA2}{eff}(x, ...)

\method{APA2}{efflist}(
  x,
  caption = "Effekte: ",
  type = NULL,
  note = "",
  output = stp25output::which_output(),
  digits = 2,
  include.fit = TRUE,
  include.n = FALSE,
  include.ci = TRUE,
  include.se = FALSE,
  ...
)

etaSquared2(x, type = 2, anova = FALSE, ...)

cohens.d(x, ...)

\method{cohens.d}{default}(x, y, ...)

\method{cohens.d}{formula}(x, data = list(), ...)
}
\arguments{
\item{x}{Objekt oder Formel}

\item{...}{}

\item{include.means}{emmeans Mittelwewrte}

\item{include.contrasts}{emmeans p_werte}

\item{type}{etaSquared2: Anova Type default ist 2}

\item{anova}{etaSquared2: Ausgabe der ANOVA Tabelle}

\item{x, y}{formel oder x, y}
}
\value{
Vector

vector
}
\description{
Measures of association
Pearson's r correlation Small 0.2,  Medium 0.5, Large 0.8
r2 coefficient of determination Small 0.04, Medium 0.25, Large 0.64
Quelle: http://www.ncbi.nlm.nih.gov/pmc/articles/PMC3444174/pdf/i1949-8357-4-3-279.pdf

emmeans: APA2_methode fuer

Eta-Quadrat Kopie von lsr::etaSquared.
Die Ergebnisse entsprechen denen von SPSS (Univariat-Partial Eta Squared)

Cohen's d and Hedges g effect size
Between groups
Cohen's d Small 0.2, Medium 0.5, Large 0.8, Very large 1.3
Odds ratio (OR) Small 1.5, Medium 2, Large 3
Relative risk or risk ratio (RR) R Small 2, Medium 3, Large 4
Cohen, J. (1988).  Statistical power analysis for the behavioral sciences (2nd ed.).  New York:Academic Press.
}
\details{
Gestolen von https://cran.r-project.org/web/packages/effsize/effsize.pdf
}
\examples{

 require(emmeans)


warp.lm <- lm(breaks ~ wool * tension, data = warpbreaks)

r1 <- emmeans (warp.lm,  ~ wool | tension)
r2 <- emmeans (warp.lm, poly ~ tension | wool)
r3 <- emmeans(warp.lm, pairwise ~ tension | wool)
APA2(r1)
APA2(r2)
APA2(r3)

# plot(r3)
# emmip(warp.lm, wool ~ tension)


#' # Effekte / Mittelwerte
fit1 <- lm(chol0 ~  ak + rrs0 + med + g, hyper)
Tabelle2(fit1, digits = 2)  # mean SD

eff <- effects::allEffects(fit1)

APA2(eff)


# etaSquared

fit1<-lm(y1~x1, anscombe)
#etaSquared2(aov (y1~x1, anscombe), anova=TRUE)
#etaSquared2(fit1, anova=TRUE)
etaSquared2(fit1 )


set.seed(45)                        ## be reproducible
 x <- rnorm(10, 10, 1)
 y <- rnorm(10, 5, 5)

cohens.d(x, y)
varanax<-Melt2(m1+m2~nr,varana , key="time", value="m")
cohens.d(m~time, varanax )

}
