% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{extract_meta}
\alias{extract_meta}
\alias{APA2.default}
\title{extract_meta}
\usage{
extract_meta(x, ...)

\method{APA2}{default}(x, ...)
}
\arguments{
\item{x}{meta-Objekt}

\item{caption, note, digits}{an Output}
}
\value{
data.frame
}
\description{
Extract results from  R package meta.
}
