% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC.R
\name{ICC}
\alias{ICC}
\alias{ICC2}
\alias{ICC.matrix}
\alias{ICC.data.frame}
\alias{ICC.formula}
\alias{AD_Index}
\alias{AD_Index2}
\alias{AD_Index.data.frame}
\alias{AD_Index.formula}
\alias{AD_Index.default}
\title{ICC Intra-Klassen-Korrelation}
\usage{
ICC(...)

ICC2(..., caption = "ICC", note = "", type = c(1, 4),
  output = which_output())

\method{ICC}{matrix}(data)

\method{ICC}{data.frame}(data)

\method{ICC}{formula}(x, data)

AD_Index(x, ...)

AD_Index2(x, data, ..., A = NULL, type = "judge",
  caption = "Interrater Agreement",
  note = "AD index (based on deviations from the item medians)",
  output = which_output())

\method{AD_Index}{data.frame}(data, ...)

\method{AD_Index}{formula}(x, data, type = "judge", ...)

\method{AD_Index}{default}(x, ...)
}
\arguments{
\item{...}{Weitere Argumente al psych missing=TRUE, alpha=.05}

\item{caption, note, output}{an Output}

\item{type}{1 bis 6 type = c("all", "ICC1", "ICC2")
c("ICC1" "ICC2", "ICC3" "ICC1k" "ICC2k"  "ICC3k")}

\item{data}{dataframe oder Matrix an psych::ICC}

\item{x}{Formula}

\item{A}{in AD_Index A=5 Anzahl der Stufen der Likertskala
na.rm normalerweise bna.rm==TRUE soll nicht geaendert werden}
}
\value{
ICC gibt ein psych -Objetht retur ICC2 einen data.frame

Ein data.frame
}
\description{
Kopie von \link{ICC} aus dem Packet \code{psych}.
 psych erwartet eine Matrix in form von:

AD_index  Hypothese Jeder Judge beurteilt das Item gleich Estimating Interrater Agreement With the
Average Deviation Index: A User’s Guide MICHAEL J. BURKE WILLIAM P. DUNLAP
Basiert auf 642 Raphaela Willitsch
wo Estimating Interrater Agreement With the Average Deviation Index: A User’s Guide MICHAEL J. BURKE
WILLIAM P. DUNLAP
die Methode beschreibt.

Estimating within-group interrater reliability with and without response bias.
By James, Lawrence R.; Demaree, Robert G.; Wolf, Gerrit
Journal of Applied Psychology, Vol 69(1), Feb 1984, 85-98.
Abstract
Presents methods for assessing agreement among the judgments made by a single group of judges on a single variable in regard to a single target. For example, the group of judges could be editorial consultants, members of an assessment center, or members of a team. The single target could be a manuscript, a lower level manager, or a team. The variable on which the target is judged could be overall publishability in the case of the manuscript, managerial potential for the lower level manager, or a team cooperativeness for the team. The methods presented are based on new procedures for estimating interrater reliability. For such situations, these procedures furnish more accurate and interpretable estimates of agreement than estimates provided by procedures commonly used to estimate agreement, consistency, or interrater reliability. The proposed methods include processes for controlling for the spurious influences of response biases (e.g., positive leniency and social desirability) on estimates of interrater reliability. (49 ref) (PsycINFO Database Record (c) 2016 APA, all rights reserved)

Hypothese Jeder Judge beurteilt das Item gleich.

rwg: oeherer Wertbessere Uebereinstimmung
}
\details{
ICC und ICC2 erlaubt die Verwendung von Formeln \code{ICC(~a+b+c+g, data)}
 wobei hier auch die Rater bzw Judge
}
\examples{
#'
#Quelle:James Demaree Woolf http://sci-hub.ac/10.1037/0021-9010.69.1.85
data<-stp25aggregate::GetData(
  "Judge   item1 item2 item3 item4 item5 item6
  1       3     4     3     2     4     3
  2       2     3     3     2     4     4
  3       4     3     4     3     4     3
  4       3     3     2     2     2     4
  5       3     2     2     4     2     3
  6       4     2     4     3     2     3
  7       2     3     2     3     3     4
  8       3     4     4     3     3     2
  9       4     2     3     4     3     2
  10      2     4     3     4     3     2")




AD_Index2(~item1+item2+item3+item4+item5+item6, data, type="item", A=4)

data2<- data.frame(t(data[,-1]))
names(data2)<-  c("J1", "J2", "J3", "J4", "J5", "J6", "J7", "J8", "J9", "J10")
AD_Index2(~J1+J2+J3+J4+J5+J6+J7+J8+J9+J10, data2, type="judge", A=4)



data<- stp25aggregate::GetData("
                               Item J1 J2 J3 J4 J5 J6 J7 J8 J9 J10
                               1  5  4  5  4  5  4  5  4  5  4
                               2  4  5  4  5  4  5  4  5  4  5
                               3  5  5  5  5  5  4  4  4  4  4
                               4  5  4  4  5  5  5  4  5  4  5")


AD_Index2(~J1+J2+J3+J4+J5+J6+J7+J8+J9+J10, data)
#Quelle:James Demaree Woolf
ICC2(~J1+J2+J3+J4+J5+J6+J7+J8+J9+J10, data)
##Head("personality-project")
#Quelle  http://www.personality-project.org/r/book/Chapter7.pdf
sf <- stp25aggregate::GetData("
                              J1 J2 J3 J4 J5 J6
                              1  1  6  2  3  6
                              2  2  7  4  1  2
                              3  3  8  6  5 10
                              4  4  9  8  2  4
                              5  5 10 10  6 12
                              6  6 11 12  4  8")
sf  # Intraclass Correlation Coefficient (ICC)
ICC2(sf)

#AD_Index2(~., data)

}
