% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Manova-Lda.R, R/zz-stp25stat.R
\docType{package}
\name{stp25stat-package}
\alias{stp25stat}
\alias{stp25stat-package}
\title{stp25stat: Functions for statistical computations}
\description{
Collection of convenient functions for common statistical computations.

Convert statistical analysis objects from R into APA-Tabls.
}
\examples{


#' ---
#' title: Daten Aufbereiten
#' author: Wolfgang Peter
#' output:
#'   pdf_document
#' keep_tex: yes
#' lang: de-DE
#' ---
#+ setup, include=FALSE
#knitr::opts_chunk$set(echo = FALSE, warnings=FALSE)
#require(stpvers)
#require(tidyverse)
#require( effects)



#' #MANOVA
require(car)
require(candisc)
DF <- stp25aggregate::GetData("C:/Users/wpete/Dropbox/3_Forschung/R-Project/stp25data/extdata/manova.sav")
DF$GROUP <-
  factor(DF$GROUP, 1:3, Cs("website", "nurse ", "video tape"))
#DF \%>\% Tabelle2(USEFUL, DIFFICULTY, IMPORTANCE, by=~GROUP )
z <- as.matrix(DF[, -1])
fit1 <- manova(z ~ DF$GROUP)

#+ results='asis'
APA2(fit1)


#+
summary(fit1)$Eigenvalues
#APA2(manova(cbind(USEFUL, DIFFICULTY, IMPORTANCE)~GROUP, DF ))


#+
fit2 <- lm(cbind(USEFUL, DIFFICULTY, IMPORTANCE) ~ GROUP, DF)
can1 <- candisc(fit2, term = "GROUP")


#+ setup-plot, include=FALSE
require(lattice)
# require(stp25plot)
# graphics.off()
reset_lattice() 
set_lattice_bw(  col = grey.colors(4, start = 0.4, end = 0.9)) 


#+ plot-1, fig.cap='Eigenwerte', fig.height=6, fig.width=6, echo=FALSE 
 plot( can1)

#+ results='asis'
APA2(can1)


#+
Anova(fit2, test = "Wilks") 


#+ results='asis'
APA2(fit2)



#+ results='asis'
APA2(Anova(fit2, test = "Pillai"))
}
\author{
\strong{Maintainer}: Wolfgang Peter \email{w.peter@statistik-peter.at}


\strong{Maintainer}: Wolfgang Peter \email{w.peter@statistik-peter.at}

}
